[Attention] Mamba attentin refactor
